requirements.txt
beautifulsoup4==4.5.1
bleach==1.5.0
Django==1.10.2
django-bleach==0.3.0
django-ckeditor==5.1.1
django-cors-headers==1.3.1
django-debug-toolbar==1.6
django-filter==1.0.1
django-htmlmin==0.10.0
django-neue-transmeta==0.8.0
djangorestframework==3.5.3
html5lib==0.9999999
httpie==0.9.6
lorem-ipsum-generator==0.3
mysqlclient==1.3.9
Pillow==3.4.2
Pygments==2.1.3
python-monkey-business==1.0.0
PyYAML==3.12
requests==2.12.1
six==1.10.0
sqlparse==0.2.2
transliterate==1.8.1
webencodings==0.5
